docs: added docs for refactor
