+++ packages/react-native-sdk/src/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './useAndroidKeepCallAliveEffect';
 export * from './useIsIosScreenshareBroadcastStarted';
 export * from './useIsInPiPMode';
 export * from './useAutoEnterPiPEffect';
+export * from './useApplyDefaultMediaStreamSettings';
