+++ packages/react-native-sdk/src/hooks/useApplyDefaultMediaStreamSettings.tsx
